UI Updated

+++ calculator.js
@@ -1,17 +1,17 @@
 function add(a,b){
-    return parseInt(a) + parseInt(b);
+    return parseFloat(a) + parseFloat(b);
 };
 
 function subtract(a,b){
-    return parseInt(a) - parseInt(b);
+    return parseFloat(a) - parseFloat(b);
 }
 
 function multiply(a,b){
-    return parseInt(a) * parseInt(b);
+    return parseFloat(a) * parseFloat(b);
 }
 
 function divide(a,b){
-    return parseInt(a) / parseInt(b);
+    return parseFloat(a) / parseFloat(b);
 }
